Refactor definitionServiceTest
Separate aggregator related tests to aggregatorTest.js
